requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
argon2-cffi==23.1.0
colorama==0.4.6
